+++ lib/rapidflow/work_item.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module Rapidflow
+module RapidFlow
   # Represents a work item flowing through the pipeline
   WorkItem = Struct.new('WorkItem', :index, :data, :error) do
     def has_error?