+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Rename module `Rapidflow` to `RapidFlow`.
+
 ## [0.1.0] - 2025.11.01
 
 ### Added